refactor:the principle of least privilege

+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -42,18 +42,14 @@ class TicketController extends ApiController
      */
     public function store(StoreTicketRequest $request)
     {
-        try{
-            //$user = User::findOrFail($request->input('data.relationships.author.data.id'));
-            //instead by line 28 in app/Http/Requests/Api/V1/StoreTicketRequest.php
-
-            //policy
-            $this->isAble('store', Ticket::class);
-
+        //$user = User::findOrFail($request->input('data.relationships.author.data.id'));
+        //instead by line 28 in app/Http/Requests/Api/V1/StoreTicketRequest.php
+        //policy
+        if($this->isAble('store', Ticket::class)){
             return new TicketResource(Ticket::create($request->mappedAttributes()));
-
-        }catch(AuthorizationException $ex){
-            return $this->error('You are not authorized to update that resource', 401);
         }
+        
+        return $this->error('You are not authorized to update that resource', 401);
     }
 
     /**
@@ -85,15 +81,15 @@ class TicketController extends ApiController
         try{
             $ticket = Ticket::findOrFail($ticket_id);
 
-            $this->isAble('update', $ticket);
+            if($this->isAble('update', $ticket)){
+                $ticket->update($request->mappedAttributes());
 
-            $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            return new TicketResource($ticket);
+            return $this->error('You are not authorized to update that resource', 401);
         }catch(ModelNotFoundException $exception){
             return $this->error('Ticket cannot be found.', 404);
-        }catch(AuthorizationException $ex){
-            return $this->error('You are not authorized to update that resource', 401);
         }
     }
 
@@ -103,16 +99,15 @@ class TicketController extends ApiController
         try{
             $ticket = Ticket::findOrFail($ticket_id);
 
-            $this->isAble('replace', $ticket);
-
-            $ticket->update($request->mappedAttributes());
+            if($this->isAble('replace', $ticket)){
+                $ticket->update($request->mappedAttributes());
 
             return new TicketResource($ticket);
+            }
+            return $this->error('You are not authorized to update that resource', 401);
         }catch(ModelNotFoundException $exception){
             return $this->error('Ticket cannot be found.', 404);
         }
-
-        
     }
 
     /**
@@ -124,11 +119,13 @@ class TicketController extends ApiController
             $ticket = Ticket::findOrFail($ticket_id);
 
             //policy
-            $this->isAble('delete', $ticket);
-
-            $ticket->delete();
-
-            return $this->ok('Ticket successfully deleted');
+            if($this->isAble('delete', $ticket)){
+                $ticket->delete();
+                
+                return $this->ok('Ticket successfully deleted');
+            }
+            
+            return $this->error('You are not authorized to delete that resource', 401);
         }catch(ModelNotFoundException $exception){
             return $this->error('Ticket cannot found.', 404);
         }
